TASK-86 [Feat]: 대시보드 상세와 멤버를 AuthHeadr로 이동
