+++ .vscode/settings.json
@@ -9,6 +9,8 @@
     "mypage",
     "pretandard",
     "pretendard",
+    "signup",
+    "tailwindcss",
     "taskify"
   ]
 }
