+++ package.json
@@ -33,7 +33,8 @@
     "react": "^18",
     "react-dom": "^18",
     "react-hook-form": "^7.52.2",
-    "tailwind-merge": "^2.4.0"
+    "tailwind-merge": "^2.4.0",
+    "tailwindcss": "^3.4.10"
   },
   "devDependencies": {
     "@types/node": "^20",
@@ -59,7 +60,6 @@
     "postcss-simple-vars": "^7.0.1",
     "prettier": "^3.3.3",
     "prettier-plugin-tailwindcss": "^0.6.5",
-    "tailwindcss": "^3.4.1",
     "typescript": "^5"
   }
 }
