+++ src/core/contexts/RootContexts.tsx
@@ -15,24 +15,20 @@ import {
 import { usePathname, useRouter } from 'next/navigation';
 
 import useApi from '@/src/lib/hooks/useApi';
-import {
-  DashboardApplicationServiceResponseDto,
-  DashboardsResponseDto,
-} from '@core/dtos/DashboardDto';
+import { DashboardsResponseDto } from '@core/dtos/DashboardDto';
+import { initialDashboard } from '@lib/constants/initialValues';
 
 import type {
   LoginRequestDto,
   LoginResponseDto,
   UserServiceResponseDto,
 } from '@core/dtos/AuthDto';
-import type { MembersResponseDto } from '@core/dtos/MembersDto';
 
 interface ContextValue {
   user: UserServiceResponseDto | undefined;
   dashboardid: string | undefined;
   dashBoardList: Partial<DashboardsResponseDto>;
-  dashBoardMembers: Partial<MembersResponseDto>;
-  dashBoardDetail: Partial<DashboardApplicationServiceResponseDto>;
+
   setDashboardid: Dispatch<SetStateAction<string | undefined>>;
   redirectDashboard: (id: number) => void;
   login: (body: LoginRequestDto) => Promise<void>;
@@ -42,47 +38,17 @@ const RootContext = createContext<ContextValue>({
   user: undefined,
   dashboardid: undefined,
   dashBoardList: {},
-  dashBoardMembers: {},
-  dashBoardDetail: {},
+
   setDashboardid: () => {},
   redirectDashboard: () => {},
   login: async () => {},
 });
 
-const initialMembers: Partial<MembersResponseDto> = {
-  members: [],
-  totalCount: 0,
-};
-
-const initialDetail: Partial<DashboardApplicationServiceResponseDto> = {
-  id: undefined,
-  title: undefined,
-  color: undefined,
-  createdAt: undefined,
-  updatedAt: undefined,
-  createdByMe: undefined,
-  userId: undefined,
-};
-
-const initialDashboard: Partial<DashboardsResponseDto> = {
-  cursorId: undefined,
-  totalCount: undefined,
-  dashboards: [],
-};
-
 export default function RootProvider({ children }: PropsWithChildren) {
   const pathname = usePathname();
   const router = useRouter();
   const [dashboardid, setDashboardid] = useState<string | undefined>(undefined);
-  const {
-    data: dashBoardMembers = initialMembers,
-    callApi: getDashBoardMembers,
-  } = useApi<MembersResponseDto>(`/members`, 'GET');
-  const { data: dashBoardDetail = initialDetail, callApi: getDashBoardDetail } =
-    useApi<DashboardApplicationServiceResponseDto>(
-      `/dashboards/${dashboardid}`,
-      'GET'
-    );
+
   const { data: loginData, callApi: postAuthLogin } = useApi<LoginResponseDto>(
     '/auth/login',
     'POST'
@@ -111,23 +77,6 @@ export default function RootProvider({ children }: PropsWithChildren) {
   );
 
   useEffect(() => {
-    const fetchMembers = async () => {
-      await getDashBoardMembers(undefined, {
-        params: {
-          dashboardId: dashboardid,
-        },
-      });
-    };
-    const fetchDetail = async () => {
-      await getDashBoardDetail(undefined);
-    };
-    if (dashboardid) {
-      fetchMembers();
-      fetchDetail();
-    }
-  }, [dashboardid, getDashBoardMembers, getDashBoardDetail, loginData]);
-
-  useEffect(() => {
     if (loginData?.accessToken) {
       localStorage.setItem('accessToken', loginData?.accessToken);
     }
@@ -145,22 +94,12 @@ export default function RootProvider({ children }: PropsWithChildren) {
     () => ({
       user,
       dashboardid,
-      dashBoardMembers,
-      dashBoardDetail,
       dashBoardList,
       setDashboardid,
       login,
       redirectDashboard,
     }),
-    [
-      user,
-      dashboardid,
-      login,
-      dashBoardMembers,
-      dashBoardDetail,
-      dashBoardList,
-      redirectDashboard,
-    ]
+    [user, dashboardid, login, dashBoardList, redirectDashboard]
   );
 
   return <RootContext.Provider value={value}>{children}</RootContext.Provider>;
